+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * Invalid type of `/DR` field of [HexaPDF::Type::AcroForm::Form]
 * Invalid ordering of types for the `/V` and `/DV` fields of
   [HexaPDF::Type::AcroForm::Field]
+* [HexaPDF::Type::AcroForm::Field#terminal_field?] to work according to the spec
 
 
 ## 0.11.9 - 2020-06-15